+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/FlexibleIpSpaceSpecifierFactory.java
@@ -4,9 +4,14 @@ import static com.google.common.base.Preconditions.checkArgument;
 
 import com.google.auto.service.AutoService;
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
+import java.util.Arrays;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.batfish.datamodel.IpWildcard;
+import org.batfish.datamodel.IpWildcardSetIpSpace;
 
 /**
  * An abstract IpSpaceSpecifierFactory that accepts the following types of inputs:
@@ -50,7 +55,33 @@ public final class FlexibleIpSpaceSpecifierFactory implements IpSpaceSpecifierFa
       return new LocationIpSpaceSpecifier(
           new FlexibleLocationSpecifierFactory().buildLocationSpecifier(matcher.group(1)));
     }
-    return new ConstantWildcardSetIpSpaceSpecifierFactory().buildIpSpaceSpecifier(input);
+    return new ConstantIpSpaceSpecifier(parseIpSpace(input));
+  }
+
+  @VisibleForTesting
+  static IpWildcardSetIpSpace parseIpSpace(@Nonnull String input) {
+    String[] strs = input.split("-|\\\\");
+
+    if (strs.length == 1) {
+      return IpWildcardSetIpSpace.builder().including(parseWildcards(strs[0])).build();
+    }
+    if (strs.length == 2) {
+      return IpWildcardSetIpSpace.builder()
+          .including(parseWildcards(strs[0]))
+          .excluding(parseWildcards(strs[1]))
+          .build();
+    }
+    throw new IllegalArgumentException(
+        "Error parsing IpWildcards: only 1 subtraction operator ('-' or '\\') allowed");
+  }
+
+  @VisibleForTesting
+  static Iterable<IpWildcard> parseWildcards(@Nonnull String wildcardsStr) {
+    String[] wildcardStrs = wildcardsStr.split(",");
+    return Arrays.stream(wildcardStrs)
+        .map(String::trim)
+        .map(IpWildcard::new)
+        .collect(ImmutableList.toImmutableList());
   }
 
   @Override
